+++ crates/karva_runner/src/lib.rs
@@ -1,7 +1,9 @@
+mod binary;
 mod collection;
 mod orchestration;
 mod partition;
 mod shutdown;
+mod worker_args;
 
 pub use orchestration::{ParallelTestConfig, RunOutput, run_parallel_tests};
 pub use shutdown::shutdown_receiver;
